+++ NEWS.md
@@ -28,3 +28,8 @@
 
 * Changed significance stars so you can choose which parameters to use them with
 
+# lavanPlot 0.5.2
+
+* Added digits option for coefficients
+
+* Changed coefs argument behavior so that it toggles coef values for covariances too